Extract import contacts UI component
